+++ relayer/CrossChainContract.sol
@@ -7,7 +7,7 @@ interface ICrossChainContract {
     function onPropose(
         uint256 taskID,
         string memory params
-    ) external returns (int16); // return true if propose check success
+    ) external returns (int256); // return true if propose check success
 
     function onCancel(uint256 taskID) external;
 
@@ -83,7 +83,7 @@ contract NativeBridge is CrossChainBridge {
             return 0;
         }
 
-        int16 success = crossChainContract.onPropose(taskID, params);
+        int256 success = crossChainContract.onPropose(taskID, params);
         if (success == 0) {
             // commit(taskID);
             return taskID; // call commit(taskID) if you get taskID
@@ -188,7 +188,7 @@ contract EvidenceManagerExample is CrossChainContract {
     function onPropose(
         uint256 taskID,
         string memory params
-    ) public override returns (int16) {
+    ) public override returns (int256) {
         // check avalaible here
         // in ERC20, you can approve money to somewhere in this function
         if (bytes(tasks[taskID]).length != 0) {
